pyproject.toml
[tool.poetry]
name = "news"
version = "0.1.0"
description = ""
authors = ["Igor <Aleshichev.git@email.ua>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.7"
django-cors-headers = "^4.9.0"
django-filter = "^25.2"
djangorestframework = "^3.16.1"
pillow = "^12.0.0"
psycopg2-binary = "^2.9.11"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.5.1"
pyjwt = "^2.10.1"
celery = "^5.5.3"
redis = "^7.0.1"
django-celery-beat = "^2.8.1"
stripe = "^13.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
